refactor: split the old perftests into files to wrap head around them
