+++ apps/omg_performance/test/omg_performance/integration/byzantine_event_test.exs
@@ -36,7 +36,7 @@ defmodule OMG.Performance.ByzantineEventsTest do
   @tag fixtures: [:perf_test, :child_chain, :omg_watcher]
   test "can provide timing of response when asking for exit data", %{perf_test: {:ok, %{destdir: destdir}}} do
     spenders = Generators.generate_users(2)
-    :ok = Performance.start_extended_perftest(100, spenders, %{randomized: false, destdir: destdir})
+    :ok = Performance.ExtendedPerftest.start(100, spenders, %{randomized: false, destdir: destdir})
     :ok = ByzantineEvents.watcher_synchronize()
     alice = Enum.at(spenders, 0)
 
@@ -47,7 +47,7 @@ defmodule OMG.Performance.ByzantineEventsTest do
   @tag fixtures: [:perf_test, :child_chain, :omg_watcher]
   test "can provide timing of status.get under many valid SEs", %{perf_test: {:ok, %{destdir: destdir}}} do
     spenders = Generators.generate_users(2)
-    :ok = Performance.start_extended_perftest(100, spenders, %{randomized: false, destdir: destdir})
+    :ok = Performance.ExtendedPerftest.start(100, spenders, %{randomized: false, destdir: destdir})
     :ok = ByzantineEvents.watcher_synchronize()
     alice = Enum.at(spenders, 0)
 
@@ -63,7 +63,7 @@ defmodule OMG.Performance.ByzantineEventsTest do
   @tag fixtures: [:perf_test, :child_chain, :omg_watcher]
   test "can provide timing of status.get under many valid/invalid SEs", %{perf_test: {:ok, %{destdir: destdir}}} do
     spenders = Generators.generate_users(2)
-    :ok = Performance.start_extended_perftest(100, spenders, %{randomized: true, destdir: destdir})
+    :ok = Performance.ExtendedPerftest.start(100, spenders, %{randomized: true, destdir: destdir})
     :ok = ByzantineEvents.watcher_synchronize()
     alice = Enum.at(spenders, 0)
 
@@ -79,7 +79,7 @@ defmodule OMG.Performance.ByzantineEventsTest do
   @tag fixtures: [:perf_test, :child_chain, :omg_watcher]
   test "can provide timing of challenging", %{perf_test: {:ok, %{destdir: destdir}}} do
     spenders = Generators.generate_users(2)
-    :ok = Performance.start_extended_perftest(100, spenders, %{randomized: true, destdir: destdir})
+    :ok = Performance.ExtendedPerftest.start(100, spenders, %{randomized: true, destdir: destdir})
     :ok = ByzantineEvents.watcher_synchronize()
     alice = Enum.at(spenders, 0)
 
@@ -101,7 +101,7 @@ defmodule OMG.Performance.ByzantineEventsTest do
   @tag fixtures: [:perf_test, :child_chain, :omg_watcher]
   test "can provide timing of status.get under many challenged SEs", %{perf_test: {:ok, %{destdir: destdir}}} do
     spenders = Generators.generate_users(2)
-    :ok = Performance.start_extended_perftest(100, spenders, %{randomized: true, destdir: destdir})
+    :ok = Performance.ExtendedPerftest.start(100, spenders, %{randomized: true, destdir: destdir})
     :ok = ByzantineEvents.watcher_synchronize()
     alice = Enum.at(spenders, 0)
